Condensed HomeNavbar and Navbar into one component
