+++ Client/src/navigation/Navbar/index.js
@@ -1,2 +1,1 @@
-export { default as HomeNavbar } from './HomeNavbar';
 export { default as Navbar } from './Navbar';
